feat: adding roms-vss-movies recipe
