+++ java-recipes/notebooks/jupyter/Dockerfile
