+++ java-recipes/notebooks/jupyter/java/kernel.json
